improved error handling by adding initial todo lookup to try catch blocks

+++ backend/routes/todos.js
@@ -65,16 +65,16 @@ router.put("/:id", async (req, res) => {
         return res.status(400).send(error.details[0].message)
     }
 
-    const todo = await Todo.findById(req.params.id)
-
-    if (!todo) {
-        console.log("Todo not found")
-        return res.status(404).send("Todo not found.")
-    }
-
-    const { name, author, uid, isComplete, date } = req.body
-
     try {
+        const todo = await Todo.findById(req.params.id)
+
+        if (!todo) {
+            console.log("Todo not found")
+            return res.status(404).send("Todo not found.")
+        }
+
+        const { name, author, uid, isComplete, date } = req.body
+
         const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, { name, author, uid, isComplete, date }, { new: true })
 
         res.send(updatedTodo)
@@ -86,13 +86,13 @@ router.put("/:id", async (req, res) => {
 });
 
 router.patch("/:id", async (req, res) => {
-    const todo = await Todo.findById(req.params.id)
-
-    if (!todo) {
-        return res.status(404).send("Todo not found.")
-    }
-
     try {
+        const todo = await Todo.findById(req.params.id)
+
+        if (!todo) {
+            return res.status(404).send("Todo not found.")
+        }
+
         const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, { isComplete: !todo.isComplete })
 
         res.send(updatedTodo)
@@ -105,6 +105,12 @@ router.patch("/:id", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
     try {
+        const todo = await Todo.findById(req.params.id)
+
+        if (!todo) {
+            return res.status(404).send("Todo not found.")
+        }
+
         const deletedTodo = await Todo.findByIdAndDelete(req.params.id)
 
         res.send(deletedTodo)
